12 commit

+++ Курсовик/Курсовик/Models/Расход.cs
@@ -11,7 +11,7 @@ namespace Курсовик.Models
         public String Название { get; set; }
         public int Приход_за_сутки { get; set; }
         public int Расход_за_сутки { get; set; }
-        public int Расход_по_норме { get; set; }
+        public float Расход_по_норме { get; set; }
         public float Коэффициент_расхода { get; set; }
         
     }
